+++ src/navbar/index.ts
@@ -1,8 +1,6 @@
 export { Navbar } from './navbar'
-export { NavbarAction } from './navbar-action'
 export { getNavbarToken } from './token'
 export type {
-  NavbarActionProps,
   NavbarProps,
   NavbarSemanticStyles,
   NavbarStyleInfo,